pyproject.toml
[project]
name = "frame-pit"
version = "0.1.0"
description = "frame pit project"
authors = [
    {name = "dobrso"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
